rename to parallel_convert
